requirement.txt
asgiref==3.8.1
Django==4.2.15
django-mongoengine==0.5.6
djongo==1.3.6
dnspython==2.6.1
mongoengine==0.28.2
pymongo==4.8.0
python-version==0.0.2
pytz==2024.1
sqlparse==0.2.4
typing_extensions==4.12.2
tzdata==2024.1
django-cors-headers==4.4.0
djangorestframework
Pillow-10.4.0
